QiVault/qieos/requirements.txt
PySimpleGUI>=5.0.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
rich>=13.7.1